backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
google-generativeai>=0.5.0